requirements.txt
Flask==2.3.3
Jinja2==3.1.2
Werkzeug==2.3.7
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
rasterio==1.3.8
GDAL==3.7.1
matplotlib==3.7.2
scikit-image==0.21.0
tqdm==4.65.0
requests>=2.31.0
gunicorn==21.2.0
pandas>=1.5.3
scikit-learn>=1.3.0